EventSpaceAlgebra

Documentation for EventSpaceAlgebra.

EventSpaceAlgebra.Distance

EventSpaceAlgebra.Distance

EventSpaceAlgebra.GeneralSpace

EventSpaceAlgebra.GeneralSpace

EventSpaceAlgebra.ValueUnit

EventSpaceAlgebra.disttype

EventSpaceAlgebra.sameunit

EventSpaceAlgebra.Distance — Type

Distance is the supertype for all types of "distance" on "GeneralSpace".

Distance derived from multiple dimensional is beyond the scope of Distance.

Each concrete struct of GeneralUnit corresponds to the only concrete struct of Distance

EventSpaceAlgebra.Distance — Method

Distance(value, unit) construct a concrete struct of Distance.

Example

Distance(5, Degree) == AngularDistance(ValueUnit(5, Degree))

# output

true

Add a new type <: Distance

"Constructor" Distance relies on disttype. Following the following steps to add new utilities:

struct NewDistance <: Distance
    vu::ValueUnit
end

struct NewDistUnit <: GeneralUnitOrOneOfItsAbstractType end

distype(::Type{<:GeneralUnitOrOneOfItsAbstractType}) = NewDistance

See also: disttype.

EventSpaceAlgebra.GeneralSpace — Type

Abstract type GeneralSpace are the supertype for all dimension/coordinate specification concrete structs of ValueUnit, such as Longitude, Latitude and EventTime.

EventSpaceAlgebra.GeneralSpace — Method

The interface for constructing any concrete type belongs to GeneralSpace.

Example

Longitude(ValueUnit(1, Degree)) == GeneralSpace(Longitude, 1, Degree)

# output

true

EventSpaceAlgebra.ValueUnit — Type

ValueUnit is a simple mutable concrete structure that holds a value and the unit::GeneralUnit of the value.

EventSpaceAlgebra.disttype — Method

Function disttype defines the one-to-one correspondance between U::GeneralUnit and T::Distance; it returns the type/constructor T.

List:

disttype(::Type{<:Angular}) = AngularDistance

disttype(::Type{<:EpochTime}) = EpochDuration

EventSpaceAlgebra.sameunit — Method

Check if two objects of ValueUnit has the same unit.