EventSpaceAlgebra

Documentation for EventSpaceAlgebra.

EventSpaceAlgebra.Coordinate

EventSpaceAlgebra.Coordinate

EventSpaceAlgebra.Distance

EventSpaceAlgebra.Distance

EventSpaceAlgebra.JulianDay

EventSpaceAlgebra.ValueUnit

Base.:+

Base.:+

Base.:+

Base.:+

Base.:-

Base.:-

Base.:-

Base.:-

EventSpaceAlgebra._create_add

EventSpaceAlgebra.disttype

EventSpaceAlgebra.sameunit

EventSpaceAlgebra.Coordinate — Type

Abstract type Coordinate are the supertype for all dimension/coordinate specification concrete structs of ValueUnit, such as Longitude, Latitude and EventTime.

EventSpaceAlgebra.Coordinate — Method

The interface for constructing any concrete type belongs to Coordinate.

Example

Longitude(ValueUnit(1, Degree)) == Coordinate(Longitude, 1, Degree)

# output

true

EventSpaceAlgebra.Distance — Type

Distance is the supertype for all types of "distance" on "Coordinate".

Distance derived from multiple dimensional is beyond the scope of Distance.

Each concrete struct of GeneralUnit corresponds to the only concrete struct of Distance

EventSpaceAlgebra.Distance — Method

Distance(value, unit) construct a concrete struct of Distance.

Example

Distance(5, Degree) == AngularDistance(ValueUnit(5, Degree))

# output

true

Add a new type <: Distance

"Constructor" Distance relies on disttype. Following the following steps to add new utilities:

struct NewDistance <: Distance
    vu::ValueUnit
end

struct NewDistUnit <: GeneralUnitOrOneOfItsAbstractType end

disttype(::Type{<:GeneralUnitOrOneOfItsAbstractType}) = NewDistance

See also: disttype.

EventSpaceAlgebra.JulianDay — Type

struct JulianDay <: EpochTime end.

Example

using Dates, EventSpaceAlgebra
evt = EventTime(2451545, JulianDay) # create an `EventTime` in absolute `JulianDay`.
DateTime(evt) # convert it to `DateTime`

# output

2000-01-01T12:00:00

EventTime(2451545, JulianDay) + Distance(3, JulianDay) |> DateTime

# output
2000-01-04T12:00:00

EventSpaceAlgebra.ValueUnit — Type

ValueUnit is a simple mutable concrete structure that holds a value and the unit::GeneralUnit of the value.

Base.:+ — Method

Otherwise, throws error CoordinateMismatch.

Base.:+ — Method

Base.:+(gs1::Coordinate, gs2::Distance) returns a copy of Coordinate of the same type as gs1, with its value being the sum of the values of gs1 and gs2.

Base.:+ — Method

Addition is commutative. Thus, +(gs2::Distance, gs1::Coordinate) falls back to +(gs1, gs2).

Base.:+ — Method

+(gs1::T, gs2::T) where {T<:Distance} performs the addition of the two Distance of the same type. It returns Distance of the same unit, with the value of their sum.

Base.:- — Method

Base.:-(gs1, gs2) do subtraction of gs1 and gs2 of the sameunit. It throws CoordinateMismatch error despite the following cases:

Base.:- — Method

Base.:-(gs1::Coordinate, gs2::Distance) returns a Distance. gs1::Coordinate can be subtracted by gs2::Distance.

Base.:- — Method

Base.:-(gs1::Distance, gs2::Distance) returns a Distance. gs1::Distance can be subtracted by gs2::Distance.

Base.:- — Method

Base.:-(gs1::T, gs2::T) where {T<:Coordinate} returns a Distance. gs1 and gs2 of the same Coordinate can be subtracted.

EventSpaceAlgebra._create_add — Method

_create_add(gs1, gs2) do sameunit check for gs1 and gs2, performs addition of their value, and return a copy of gs1 with the value of gs1 the sum of value of the input arguments.

EventSpaceAlgebra.disttype — Method

Function disttype defines the one-to-one correspondance between U::GeneralUnit and T::Distance; it returns the type/constructor T. disttype is essential for Distance to construct an Distance struct.

List:

disttype(::Type{<:Angular}) = AngularDistance

disttype(::Type{<:EpochTime}) = EpochDuration

EventSpaceAlgebra.sameunit — Method

Check if two objects of ValueUnit has the same unit.